extract YOUTUBE_PLAYER as openwhydYouTubeExtractor, in its own TS file
